GETP-168 rename: 피플 좋아요 애그리거트를 like 패키지로 이동
